Simulator embryo
Adding a first attempt of simulating the game engine for doing tests. AntMap.pm
contain a map parser and AntBot.pm is responsible for the IPC with a Bot.
one_move.pl is updating the bot with current information and let it do one move.

What is left to do here is the updating of the map. I guess I have to rework
the internal representation of the map to be able to do updates easier. But it
might be possible to do some ugly solution as well with the existing structure.

Next step is to add script that can simulate more than just one move.