Boiding
Trying to get some boiding behaviour. What I want: Find the ant closest to the
enemy ant, and make that ant sergeant. Send a squad of ants to the sergeant. (A
squad is 5-12 ants).

Updates:
Bot.cpp - the boiding code.
Ant.{h,c}pp - A desire function wantToGoTo(), which only set the destination
desire instead of goTo(). This is for avoiding collisions.
boid.map - updated for running with simulator.
stimulate.pl - wrap-around-bug fixed.